Use json kinds for character strings
When I started this project, I was using a built-from-source
version but at some point, I installed the homebrew version.

My build was still using the built-from-source version but at
some point I cleared the build and retried to build and it

      206 |             call jcore%get(jeinfo, 'name', ei%name, found)
          |                      1
    Error: Found no matching specific binding for the call to the
    GENERIC 'get' at (1)

and this time it was using the homebrew version which is
8.3.0 as well.

I tried going back to my built-from-source version and did not
get those errors.

I looked at the source code and noticed they were using

    character(kind=CK...)

and I remembered seeing that in the documentation.  I made the
changes in this commit and now it works with both version of
json-fortran.

Checking the brew recipe json-fortran.rb, it does

    -DENABLE_UNICODE:BOOL=TRUE

in the CMake command so it is quite probable that when I'm using
my built-from-source, the kind CK is equivalent to the default
kind so it worked but when using the homebrew version, it's
different.